src/modules/ebd/ebd.controller.spec.ts
// src/ebd/ebd.e2e-spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { INestApplication } from '@nestjs/common';
import * as request from 'supertest';
import { TypeOrmModule } from '@nestjs/typeorm';
import { EbdModule } from './ebd.module';

jest.setTimeout(10000);

describe('EbdController (e2e)', () => {
  let app: INestApplication;

  beforeAll(async () => {
    const moduleFixture: TestingModule = await Test.createTestingModule({
      imports: [
        TypeOrmModule.forRoot({
          type: 'postgres',
          host: 'localhost',
          port: 5432,
          username: 'postgres',
          password: 'admin',
          database: 'cbatos29',
          entities: [__dirname + '/../**/*.entity{.ts,.js}'],
          synchronize: true,
        }),
        EbdModule,
      ],
    }).compile();

    app = moduleFixture.createNestApplication();
    await app.init();
  });

  afterAll(async () => {
    await app.close();
  });

  it('/ebd (POST)', async () => {
    return request(app.getHttpServer())
      .post('/ebd')
      .send({
        name: 'New EBD',
        classId: 1,
      })
      .expect(201)
      .then((response) => {
        expect(response.body).toHaveProperty('id');
        expect(response.body.name).toBe('New EBD');
      });
  });

  it('/ebd/:ebdId (GET)', async () => {
    return request(app.getHttpServer())
      .get('/ebd/1')
      .expect(200)
      .then((response) => {
        expect(response.body).toHaveProperty('id');
        expect(response.body.name).toBeDefined();
      });
  });
});
